+++ Clustering/connect_clusters.py
@@ -12,7 +12,7 @@ import transformers
 from sentence_transformers import SentenceTransformer
 from sklearn.metrics.pairwise import cosine_similarity
 
-from BranchingEntropy import make_event_name
+from branching_entropy import make_event_name
 
 """### Load data"""
 def load_data(date: str):
@@ -128,7 +128,7 @@ def visualize_iso_con(today: str):
     return total_result
 
 ### mysql 연결 example
-from dataset import MysqlConnection # import 구문 있는 곳(상단)에 넣으면 됨
+from database import MysqlConnection # import 구문 있는 곳(상단)에 넣으면 됨
 db_connection = MysqlConnection()
 conn = db_connection.connection
 cursor = conn.cursor()
